[T-1] First commit, basics done.
### Summary

First commit and last commit for the moment on an exemple implementation of UVM.
This implementation is made for C#. It has differente features and is working on it self.

It shows through flags, how to pass arguments to the kernel.
It also shows through its configuration, how to augment UVM for your specific needs.
Most of the work is then done in the BLL as the engine and services are just sugar for devs.

For any project that follow the same template that the csproj of UVM4Cs use,
should have a plug and play working solution for their automatic package generation and upgrade.

### Major

...

### Minor

...

###  Patch

...

### Note

The teams will not add anything to that solution, so do not expect any futher features.
This has just been a project to serve as an exemple for how powerfull UVM is.
We will not documentate the code futher and will not do support.

Hope that this project will still have your attention.

Best regard.

The team.